Forgot a comma and cast

+++ IguideME.Web/Services/Workers/QuizWorker.cs
@@ -64,7 +64,7 @@ namespace IguideME.Web.Services.Workers
 					entry.ID,
 					DatabaseManager.Instance.GetUserID(this._courseID, sub.UserID),
 					// sub.Score ?? 0,
-					(sub.Score ?? 0) / quiz.PointsPossible * 100 //, Should change to this.
+					(double)((sub.Score ?? 0) / quiz.PointsPossible * 100), //, Should change to this.
 					"",
 					this._hashCode
 				);
